Extract reconciliation calibration sub-helpers into a partial
Increment 2k of the endpoint-monolith decomposition. Moves the 10 reconciliation
calibration summary sub-helpers (CalculateAutoMatchRate, CalculateT0ClosureRate,
BuildCalibrationProfileSummary, DetermineReconciliationCalibrationStatus,
BuildReconciliationCalibrationSummaryText, HasMissingCalibrationMetadata,
RequiresCalibrationSignoff, IsTerminalCalibrationSignoff, IsSignedOff,
NormalizeCalibrationValue) out of
src/Meridian.Ui.Shared/Endpoints/WorkstationEndpoints.cs into a new sibling
partial WorkstationEndpoints.ReconciliationCalibration.cs.

Behavior-preserving verbatim relocation: the cluster is a pure closed-graph leaf
(zero references outside WorkstationEndpoints.cs). Its shared parent
BuildReconciliationCalibrationSummary — called from both core and
WorkstationEndpoints.Reconciliation.cs — stays in core and reaches the moved
helpers across the partial; the cluster makes no outbound calls to other core
helpers.

Core WorkstationEndpoints.cs drops 5,566 -> 5,452 lines; ratchet baseline
retightened.
